Moved CompanionMenu
